requirements.txt
asgiref==3.2.7
Django==3.0.6
future==0.18.2
python-telegram-bot==12.7
pytz==2020.1
sqlparse==0.3.1
